rolled back numeric change

+++ backend/db/migrations/20231128160717-create-event.js
@@ -49,7 +49,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.NUMERIC,
         allowNull: false
       },
       startDate: {
